==> sunrui/zhihu_final/3_key_word_extract.ipynb <==
%load_ext autotime

import pickle

import pandas as pd
from tqdm import tqdm

tqdm.pandas()

with open('./pkl/question_info.pkl', 'rb') as f:
    question_info = pickle.load(f)

with open('./pkl/answer_info.pkl', 'rb') as f:
    answer_info = pickle.load(f)

answer_info['content_w_series'] = answer_info['content_w_series'].progress_apply(
    lambda x: [str(i) for i in x])
question_info['title_w_series'] = question_info['title_w_series'].progress_apply(
    lambda x: [str(i) for i in x])
question_info['desc_w_series'] = question_info['desc_w_series'].progress_apply(
    lambda x: [str(i) for i in x])

question_word_titles = question_info['title_w_series'].values.tolist()
question_word_desc = question_info['desc_w_series'].values.tolist()
answer_word_content = answer_info['content_w_series'].values.tolist()

from gensim.corpora import Dictionary
from gensim.models import TfidfModel

dictionary = Dictionary(question_word_titles + question_word_desc + answer_word_content)

tf_idf_model = TfidfModel(dictionary=dictionary)

corpus_q_title = [dictionary.doc2bow(text) for text in question_word_titles]
corpus_q_desc = [dictionary.doc2bow(text) for text in question_word_desc]
corpus_answer = [dictionary.doc2bow(text) for text in answer_word_content]

id2token = {v: k for k, v in dictionary.token2id.items()}

tf_idf_model[corpus_q_title[0]]

def generate_key_words(corpus, top=5):
    key_words_result = []
    for item in tqdm(corpus):
        gensim_dict = {}
        for w, n in tf_idf_model[item]:
            gensim_dict[id2token.get(w)] = n
        key_words = sorted(gensim_dict.items(), key=lambda x: x[1], reverse=True)[:top]
        key_words_result.append([int(k) for k, v in key_words])
    return key_words_result


key_words_q_title_list = generate_key_words(corpus_q_title)

key_words_q_desc_list = generate_key_words(corpus_q_desc)
key_words_answer_list = generate_key_words(corpus_answer)

print(key_words_q_title_list[0])
print(key_words_q_desc_list[0])
print(key_words_answer_list[0])

key_words_q_desc_df = pd.DataFrame()
key_words_q_desc_df['key_words_q_desc'] = key_words_q_desc_list
key_words_q_desc_df.to_hdf('./key_words_q_desc_gensim.h5', key='data')

key_words_answer_df = pd.DataFrame()
key_words_answer_df['key_words_answer'] = key_words_answer_list
key_words_answer_df.to_hdf('./key_words_answer_gensim.h5', key='data')

key_words_q_title_df = pd.DataFrame()
key_words_q_title_df['key_words_q_title'] = key_words_q_title_list
key_words_q_title_df.to_hdf('./key_words_q_title_gensim.h5', key='data')

display(key_words_q_desc_df.head())
display(key_words_answer_df.head())
display(key_words_q_title_df.head())

import pandas as pd

key_words_q_desc = pd.read_hdf('./key_words_q_desc_gensim.h5', key='data')

key_words_q_desc.head()
